Plots now have some basic control to prevent crowding.
I found it difficult to get the xaxis of my plots to
have labels that didn't get crowded when I had
multiple figures on one page. There is now some basic
mechanisim, in which we limit the number of labels
depending on how many plots we have on a page.

Also changed the save figure name so that it won't
automatically convert it to filename001.png if you
only have one plot.

Some other minor things, including updating of the
document, though still lots of formatting to do.